pyproject.toml
[project]
name = "idh-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow~=2.9.0",
    "azure-core>=1.36.0",
    "azure-storage-blob>=12.27.1",
    "black>=25.12.0",
    "duckdb>=1.4.3",
    "google-cloud-bigquery>=3.39.0",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.3.251201",
    "python-dotenv>=1.2.1",
]
